make
git-svn-id: https://daisy-model.googlecode.com/svn/trunk@2300 5f5bf4bb-3343-0410-8278-4fa8eb3319dc

+++ groundwater_deep.C
@@ -33,7 +33,7 @@ class GroundwaterDeep : public Groundwater
 public:
   bottom_t bottom_type() const
   { return free_drainage; }
-  double q_bottom () const
+  double q_bottom (size_t) const
   { daisy_notreached (); }
 
   // Simulation.
