+++ groundwater_lysimeter.C
@@ -35,7 +35,7 @@ class GroundwaterLysimeter : public Groundwater
 public:
   bottom_t bottom_type () const
   { return lysimeter; }
-  double q_bottom () const
+  double q_bottom (size_t) const
   { daisy_notreached (); }
   bool is_lysimeter () const
   { return true; }
